Article validation fix (#121)
* Added featured to articles
Applied a fix for event staging example link and atricle site and category fields.

* Article site altering needs to allow for new articles

* fixes for Jcrop delete on non photo, author index and validation

+++ app/assets/javascripts/components/crop_preview.js.coffee
@@ -64,4 +64,5 @@ $(document).ready ->
     # if a user hits the 'change' button, you want to destroy the old JCrop
     $(".btn-file input:file").click ->
       JcropAPI = $('#cropbox').data('Jcrop');
-      JcropAPI.destroy();
+      if (JcropAPI != undefined)
+        JcropAPI.destroy();
